Safari-609.1.11.1/LayoutTests/http/tests/media/hls/hls-webvtt-tracks-expected.txt


EVENT(addtrack)
EXPECTED (video.textTracks.length == '1') OK
RUN(video.textTracks[0].mode = 'showing')
EVENT(cuechange)
EXPECTED (video.textTracks[0].activeCues.length == '1') OK
EXPECTED (video.textTracks[0].activeCues[0].text == 'Subtitle Test') OK
END OF TEST
